+++ src/utils/noop.ts
@@ -1,0 +1,4 @@
+export function noop(): void {}
+export function noopPromise(): Promise<void> {
+    return Promise.resolve();
+}
